Split search palette state and components
